fixed inconsitent coverage across the different configurations and added all types of badges to the gh-pages export for use in readme
